only the creator of the post can delete and update their post

+++ src/routes/user-routes.js
@@ -3,11 +3,10 @@
 "use strict";
 
 const express = require("express");
-const { posting } = require("../models/index.js");
 const bearerHandler = require("../auth/bearer.js");
 const aclHandler = require("../auth/acl.js");
 const socketService = require("../socket.js");
-
+const {posting} = require('../models/index');
 const router = express.Router();
 
 router.post(
@@ -32,7 +31,7 @@ router.delete(
 async function handleCreatePosting(req, res) {
   try {
     let obj = req.body;
-    console.log("****************", req.user.dataValues.username);
+    // console.log("****************", req.user.dataValues.username);
     let sellerObj = { seller: `${req.user.dataValues.username}`, ...obj };
     let newPosting = await posting.create(sellerObj);
     socketService("update", {
@@ -49,9 +48,20 @@ async function handleUpdatePosting(req, res) {
   try {
     const id = req.params.id;
     const obj = req.body;
-    console.log("**********************", id, obj);
-    let updatedPosting = await posting.update(obj, { where: { id } });
-    res.status(200).json(updatedPosting);
+    let updator = req.user.dataValues.username;
+    let oldPost = await posting.findOne({where: {id}});
+    let poster = oldPost.dataValues.seller;
+
+    // console.log("**********************");
+    // console.log('oldPost', oldPost);
+    // console.log('updator', updator);
+    // console.log('poster', poster);
+    if (updator === poster) {
+      let updatedPosting = await posting.update(obj, { where: { id } });
+      res.status(200).json(updatedPosting);
+    } else {
+      throw new Error('Invalid user');
+    }
   } catch (err) {
     console.error(err);
   }
@@ -60,8 +70,15 @@ async function handleUpdatePosting(req, res) {
 async function handleDeletePosting(req, res) {
   try {
     let id = req.params.id;
-    let deletedPosting = await posting.destroy({ where: { id } });
-    res.status(200).json(deletedPosting);
+    let deleter = req.user.dataValues.username;
+    let oldPost = await posting.findOne({where: {id}});
+    let poster = oldPost.dataValues.seller;
+    if (deleter === poster) {
+      let deletedPosting = await posting.destroy({ where: { id } });
+      res.status(200).json(deletedPosting);
+    } else {
+      throw new Error('Invalid user');
+    }
   } catch (err) {
     console.error(err);
   }
